This avoids allocating an array each time `ParameterFilter#filter` is called, and avoids a few redundant operations when building up the full key.  This yields a small but reproducible performance increase:

<details>
  <summary><strong>Benchmark script</strong></summary>

  ```ruby
  # frozen_string_literal: true
  require "benchmark/ips"
  require "benchmark/memory"
  require "active_support"
  require "active_support/parameter_filter"

  ootb = [:passw, :secret, :token, :_key, :crypt, :salt, :certificate, :otp, :ssn]
  mixed = [:passw, "secret", /token/, :crypt, "salt", /certificate/, "user.otp", /user\.ssn/, proc {}]

  params = {
    "user" => {
      "name" => :name,
      "email" => :email,
      "password" => :password,
      "ssn" => :ssn,
      "locations" => [
        { "city" => :city, "country" => :country },
        { "city" => :city, "country" => :country },
      ],
    }
  }

  Benchmark.ips do |x|
    x.report("ootb") do
      ActiveSupport::ParameterFilter.new(ootb).filter(params)
    end

    x.report("mixed") do
      ActiveSupport::ParameterFilter.new(mixed).filter(params)
    end
  end
  ```
</details>

**Before**

  ```
  Warming up --------------------------------------
                  ootb     2.003k i/100ms
                 mixed     1.362k i/100ms
  Calculating -------------------------------------
                  ootb     19.975k (± 0.9%) i/s -    100.150k in   5.014277s
                 mixed     13.627k (± 0.9%) i/s -     69.462k in   5.097741s
  ```

**After**

  ```
  Warming up --------------------------------------
                  ootb     2.032k i/100ms
                 mixed     1.521k i/100ms
  Calculating -------------------------------------
                  ootb     20.315k (± 1.3%) i/s -    101.600k in   5.001939s
                 mixed     15.142k (± 1.2%) i/s -     76.050k in   5.023077s
  ```